24.12.06

+++ test3/src/main/java/net/scit/test3/repository/FitnessRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import net.scit.test3.entity.FitnessEntity;
 
-public interface FitnessRepository extends JpaRepository<FitnessEntity, Integer>{
+public interface FitnessRepository extends JpaRepository<FitnessEntity, Long>{
 
 }
